Minor docs fixes
