+++ examples/02-ui-components/05-side-menu-drag-handle-items/App.tsx
@@ -5,11 +5,12 @@ import {
   RemoveBlockItem,
   SideMenu,
   SideMenuController,
-  useComponentsContext,
   useCreateBlockNote,
 } from "@blocknote/react";
 import { BlockNoteView } from "@blocknote/mantine";
 import "@blocknote/mantine/style.css";
+
+import { ResetBlockTypeItem } from "./ResetBlockTypeItem";
 
 export default function App() {
   // Creates a new editor instance.
@@ -34,8 +35,6 @@ export default function App() {
     ],
   });
 
-  const Components = useComponentsContext()!;
-
   // Renders the editor instance.
   return (
     <BlockNoteView editor={editor} sideMenu={false}>
@@ -48,12 +47,7 @@ export default function App() {
                 <RemoveBlockItem {...props}>Delete</RemoveBlockItem>
                 <BlockColorsItem {...props}>Colors</BlockColorsItem>
                 {/* Item which resets the hovered block's type. */}
-                <Components.Generic.Menu.Item
-                  onClick={() => {
-                    editor.updateBlock(props.block, { type: "paragraph" });
-                  }}>
-                  Reset Type
-                </Components.Generic.Menu.Item>
+                <ResetBlockTypeItem {...props}>Reset Type</ResetBlockTypeItem>
               </DragHandleMenu>
             )}
           />
